First table column should not be hideable
