+++ front/src/modules/ui/table/components/EntityTableHeaderOptions.tsx
@@ -1,15 +1,10 @@
 import styled from '@emotion/styled';
 
 import { DropdownButton } from '@/ui/dropdown/components/DropdownButton';
-import { StyledDropdownMenu } from '@/ui/dropdown/components/StyledDropdownMenu';
-import { StyledDropdownMenuItemsContainer } from '@/ui/dropdown/components/StyledDropdownMenuItemsContainer';
-import { useDropdownButton } from '@/ui/dropdown/hooks/useDropdownButton';
-import { ViewFieldMetadata } from '@/ui/editable-field/types/ViewField';
-import { IconArrowLeft, IconArrowRight, IconEyeOff } from '@/ui/icon';
-import { MenuItem } from '@/ui/menu-item/components/MenuItem';
-
-import { useTableColumns } from '../hooks/useTableColumns';
-import { ColumnDefinition } from '../types/ColumnDefinition';
+import {
+  DropdownComponents,
+  EntityTableHeaderOptionsProps,
+} from '@/ui/table/components/DropdownComponents';
 
 const StyledDropdownContainer = styled.div`
   left: 0px;
@@ -17,70 +12,24 @@ const StyledDropdownContainer = styled.div`
   top: 32px;
   z-index: 1;
 `;
-type EntityTableHeaderOptionsProps = {
-  column: ColumnDefinition<ViewFieldMetadata>;
-  isFirstColumn: boolean;
-  isLastColumn: boolean;
-};
+
 export const EntityTableHeaderOptions = ({
   column,
   isFirstColumn,
   isLastColumn,
+  primaryColumnKey,
 }: EntityTableHeaderOptionsProps) => {
-  const {
-    handleColumnVisibilityChange,
-    handleColumnLeftMove,
-    handleColumnRightMove,
-  } = useTableColumns();
-
-  const { closeDropdownButton } = useDropdownButton({
-    dropdownId: column.key + '-header',
-  });
-
-  const handleColumnMoveLeft = () => {
-    closeDropdownButton();
-    if (isFirstColumn) {
-      return;
-    }
-    handleColumnLeftMove(column);
-  };
-
-  const handleColumnMoveRight = () => {
-    closeDropdownButton();
-    if (isLastColumn) {
-      return;
-    }
-    handleColumnRightMove(column);
-  };
-
-  const handleColumnVisibility = () => {
-    handleColumnVisibilityChange(column);
-  };
-
   return (
     <StyledDropdownContainer>
       <DropdownButton
         dropdownId={column.key + '-header'}
         dropdownComponents={
-          <StyledDropdownMenu>
-            <StyledDropdownMenuItemsContainer>
-              <MenuItem
-                LeftIcon={IconArrowLeft}
-                onClick={handleColumnMoveLeft}
-                text="Move left"
-              />
-              <MenuItem
-                LeftIcon={IconArrowRight}
-                onClick={handleColumnMoveRight}
-                text="Move right"
-              />
-              <MenuItem
-                LeftIcon={IconEyeOff}
-                onClick={handleColumnVisibility}
-                text="Hide"
-              />
-            </StyledDropdownMenuItemsContainer>
-          </StyledDropdownMenu>
+          <DropdownComponents
+            column={column}
+            isFirstColumn={isFirstColumn}
+            isLastColumn={isLastColumn}
+            primaryColumnKey={primaryColumnKey}
+          />
         }
         dropdownHotkeyScope={{ scope: column.key + '-header' }}
       />
