+++ front/src/modules/ui/table/options/components/TableOptionsDropdownContent.tsx
@@ -95,7 +95,7 @@ export const TableOptionsDropdownContent = () => {
 
   const handleReorderField: OnDragEndResponder = useCallback(
     (result) => {
-      if (!result.destination) {
+      if (!result.destination || result.destination.index === 0) {
         return;
       }
